done business logic

+++ lab3/src/VoteMessage.java
@@ -1,9 +1,11 @@
+import java.math.BigInteger;
+
 public class VoteMessage {
-    public final int id;
-    public final int regId;
+    public final BigInteger id;
+    public final BigInteger regId;
     public final Ballot ballot;
 
-    public VoteMessage(int id, int regId, Ballot ballot) {
+    public VoteMessage(BigInteger id, BigInteger regId, Ballot ballot) {
         this.id = id;
         this.regId = regId;
         this.ballot = ballot;
